removeing sql

+++ NSSServices/Controllers/RegressionRegionsController.cs
@@ -25,6 +25,7 @@ using NSSDB.Resources;
 using NSSAgent;
 using System.Threading.Tasks;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace NSSServices.Controllers
 {
@@ -36,16 +37,33 @@ namespace NSSServices.Controllers
 
         #region METHOD
         [HttpGet]
-        public async Task<IActionResult> Get()
+        [HttpGet("/Regions/{regions}/[controller]")]
+        public async Task<IActionResult> GetRegressionRegions(string regions = "", [FromQuery] string statisticgroups = "", [FromQuery] string regressiontypes = "")
         {
+            IQueryable<RegressionRegion> entities = null;
+            List<string> RegionList = null;
+            List<string> statisticgroupList = null;
+            List<string> regressiontypeList = null;
             try
             {
-                return Ok(agent.GetRegressionRegions());  
+                if (String.IsNullOrEmpty(regions) &&
+                    string.IsNullOrEmpty(statisticgroups) && string.IsNullOrEmpty(regressiontypes))
+                { entities = agent.GetRegressionRegions(); }
+                else
+                {
+                    RegionList = parse(regions);
+                    statisticgroupList = parse(statisticgroups);
+                    regressiontypeList = parse(regressiontypes);
+
+                    entities = agent.GetRegressionRegions(RegionList, statisticgroupList, regressiontypeList);
+                }
+
+                return Ok(entities);
             }
             catch (Exception ex)
             {
                 return await HandleExceptionAsync(ex);
-            }      
+            }
         }
 
         [HttpGet("{id}")]
@@ -63,30 +81,25 @@ namespace NSSServices.Controllers
             }
         }
 
-        [HttpGet("Regions/{region}/[controller]")]
-        //[HttpGet("[controller]?region={region}")]
-        public async Task<IActionResult> GetRegressionRegions(int region, [FromQuery] string statisticgroups = "", [FromQuery] string regressiontypes = "")
+        [HttpPost]
+        [HttpPost("/Regions/{region}/[controller]")]
+        [Authorize(Policy = "CanModify")]
+        public async Task<IActionResult> Post([FromQuery]string region, [FromBody]RegressionRegion entity)
         {
             try
             {
-                //if (id < 0) return new BadRequestResult(); // This returns HTTP 404
+                //entity must include citation
+                if (!isValid(entity)) return BadRequest(); // This returns HTTP 404
+                Region regionEntity = agent.GetRegionByIDOrCode(region);
+                if (regionEntity == null) return BadRequest($"No region exists with {region} identifier.");              
+                if (!IsAuthorizedToEdit(regionEntity)) return Unauthorized();
 
-                return Ok();
-            }
-            catch (Exception ex)
-            {
-                return await HandleExceptionAsync(ex);
-            }
-        }
-
-        [HttpPost][Authorize(Policy = "CanModify")]
-        public async Task<IActionResult> Post([FromBody]RegressionRegion entity)
-        {
-            try
-            {
-#warning check if logged in user allowed to modify based on regionManager
-                if (!isValid(entity)) return new BadRequestResult(); // This returns HTTP 404
-                return Ok(await agent.Add(entity));
+                RegionRegressionRegion Addeditem = await agent.Add(new RegionRegressionRegion
+                                                                        {
+                                                                            RegionID = regionEntity.ID,
+                                                                            RegressionRegion = entity
+                                                                        });
+                return Ok(Addeditem.RegressionRegion);
             }
             catch (Exception ex)
             {
@@ -94,17 +107,23 @@ namespace NSSServices.Controllers
             }            
         }
 
-        [HttpPost][Authorize(Policy = "CanModify")]
-        [Route("Batch")]
-        public async Task<IActionResult> Batch([FromBody]List<RegressionRegion> entities)
+        [HttpPost("[action]")]
+        [Authorize(Policy = "CanModify")]
+        [HttpPost("/Regions/{region}/[controller]/[action]")]
+        public async Task<IActionResult> Batch([FromQuery]string region,[FromBody]List<RegressionRegion> entities)
         {
+            List<RegionRegressionRegion> rrr = new List<RegionRegressionRegion>();
             try
             {
-#warning check if logged in user allowed to modify based on regionManager
-
                 if (!isValid(entities)) return new BadRequestObjectResult("Object is invalid");
+                Region regionEntity = agent.GetRegionByIDOrCode(region);
+                if (regionEntity == null) return BadRequest($"No region exists with {region} identifier.");
+                if (!IsAuthorizedToEdit(regionEntity)) return Unauthorized();
+                                
+                entities.ForEach(rr=>rrr.Add(new RegionRegressionRegion() { RegionID = regionEntity.ID, RegressionRegion = rr }));
+                var results = await agent.Add(rrr);
 
-                return Ok(await agent.Add(entities));
+                return Ok( results.Select(r=>r.RegressionRegion));
             }
             catch (Exception ex)
             {
@@ -117,16 +136,19 @@ namespace NSSServices.Controllers
         {
             try
             {
-#warning check if logged in user allowed to modify based on regionManager
 
                 if (id < 0 || !isValid(entity)) return new BadRequestResult(); // This returns HTTP 404
+
+                RegressionRegion rr = await agent.GetRegressionRegion(id);
+                if (rr == null) return BadRequest($"No regression region exists with {id} identifier.");
+                if (!IsAuthorizedToEdit(rr)) return Unauthorized();
+
                 return Ok(await agent.Update(id,entity));
             }
             catch (Exception ex)
             {
                 return await HandleExceptionAsync(ex);
             }
-
         }        
 
         [HttpDelete("{id}")][Authorize(Policy = "CanModify")]
@@ -134,8 +156,11 @@ namespace NSSServices.Controllers
         {
             try
             {
-#warning check if logged in user allowed to modify based on regionManager
                 if (id < 1) return new BadRequestResult();
+                RegressionRegion rr = await agent.GetRegressionRegion(id);
+                if (rr == null) return BadRequest($"No regression region exists with {id} identifier.");
+                if (!IsAuthorizedToEdit(rr)) return Unauthorized();
+
                 await agent.DeleteRegressionRegion(id);
 
                 return Ok();
@@ -145,8 +170,6 @@ namespace NSSServices.Controllers
                 return await HandleExceptionAsync(ex);
             }
         }
-        #endregion
-        #region HELPER METHODS
         #endregion
     }
 }
